+++ src/client/components/AccountEditPanel/AccountEditPanel.js
@@ -258,7 +258,7 @@ Template.AccountEditPanel.events({
         //console.debug( 'item', item );
         // we cannot call here AccountHub.ahClass.preferredLabel() as this later requires an id - so compute something not too far of that
         //  must have at least one of these
-        const label = ( item.emails && item.emails[0]?.address ) || item.username || ( item.usernames && item.usernames[0]?.username ) || item._id;
+        const label = ( item.emails && item.emails[0]?.address ) || item.username || ( item.usernames && item.usernames[0]?.username ) || item._id || pwixI18n.label( I18N, 'edit.new_label' );
         // when creating a new account, we let the user create several by reusing the same modal
         const updateRoles = async function( user ){
             if( instance.AM.amInstance.get().haveRoles()){